machines: Move buildFailHandler to oVirt code since it's only used there now
VM_ACTION_FAILED has a reducer in machines code, but oVirt code does not use
the part of the state that this reducer was updating so it can be safely
skipped.
