+++ pkg/machines/constants/store-action-types.js
@@ -39,4 +39,3 @@ export const UPDATE_OS_INFO_LIST = "UPDATE_OS_INFO_LIST";
 export const UPDATE_STORAGE_VOLUMES = "UPDATE_STORAGE_VOLUMES";
 export const UPDATE_UI_VM = "UPDATE_UI_VM";
 export const UPDATE_VM = "UPDATE_VM";
-export const VM_ACTION_FAILED = "VM_ACTION_FAILED";
